pyproject.toml
[tool.poetry]
name = "smartjob"
version = "0.1.0"
description = ""
authors = ["Fabien MARTY <fabien.marty@botify.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
google-cloud-run = "^0.10.8"
google-cloud-aiplatform = "^1.60.0"
setuptools = "^72.1.0"
stlog = "^0.2.0"
rich = "^13.7.1"


[tool.poetry.group.dev.dependencies]
ruff = "^0.5.5"
mypy = "^1.11.1"
import-linter = "^2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
